+++ src/main/kotlin/no/nav/familie/ef/sak/regler/Regel.kt
@@ -1,10 +1,5 @@
 package no.nav.familie.ef.sak.regler
 
-import com.fasterxml.jackson.annotation.JsonIgnore
-import com.fasterxml.jackson.annotation.JsonProperty
-import no.nav.familie.kontrakter.felles.objectMapper
-import kotlin.reflect.KClass
-
 interface RegelId {
 
     val id: String
@@ -15,10 +10,11 @@ interface RegelIdMedBeskrivelse : RegelId {
     val beskrivelse: String
 }
 
+/*
 enum class RegelType {
     REGEL_STEG,
     RESULTAT
-}
+}*/
 
 enum class Begrunnelse {
     //UTEN,
@@ -31,26 +27,18 @@ enum class Resultat {
     IKKE_OPPFYLT
 }
 
+enum class JaNei {
+    JA,
+    NEI
+}
+
 interface RegelFlyt {
 
     //val type: RegelType
 }
 
+//TODO burde mappe til JA/NEI?
 interface Årsak {
 
-    val resultat: Resultat
-}
-
-class RegelSteg(val regelId: RegelId,
-                val hvisJa: RegelId = ResultatRegel.OPPFYLT,
-                val hvisNei: RegelId = ResultatRegel.IKKE_OPPFYLT,
-                val hvisJaBegrunnelse: Begrunnelse? = null,
-                val hvisNeiBegrunnelse: Begrunnelse? = null,
-                @JsonIgnore
-                val årsaker: KClass<out Årsak>? = null) : RegelFlyt {
-
-    @JsonProperty("årsaker")
-    val årsakMap = årsaker?.let { årsak -> årsak.java.enumConstants.map { it to it.resultat }.toMap() }
-
-    //override val type = RegelType.REGEL_STEG
+    val mapping: JaNei
 }
