+++ src/main/kotlin/no/nav/familie/ef/sak/regler/vilkårsregel/ForutgåendeMedlemskap.kt
@@ -3,12 +3,12 @@ package no.nav.familie.ef.sak.regler
 private enum class ForutgåendeMedlemskapRegel(override val id: String,
                                               override val beskrivelse: String) : RegelIdMedBeskrivelse {
 
-    SØKER_MEDLEM_I_FOLKETRYGDEN("M1", "Har søker vært medlem i folketrygden i de siste 5 årene?"),
-    UNNTAK("M2", "Er unntak fra hovedreglen oppfylt?")
+    SØKER_MEDLEM_I_FOLKETRYGDEN("F1", "Har søker vært medlem i folketrygden i de siste 5 årene?"),
+    UNNTAK("F2", "Er unntak fra hovedreglen oppfylt?")
 }
 
 @Suppress("unused")
-enum class MedlemskapUnntakÅrsak(override val resultat: Resultat = Resultat.OPPFYLT) : Årsak {
+enum class MedlemskapUnntakÅrsak(override val mapping: JaNei = JaNei.JA) : Årsak {
 
     MEDLEM_MER_ENN_5_ÅR_AVBRUDD_MINDRE_ENN_10_ÅR,
     MEDLEM_MER_ENN_7_ÅR_AVBRUDD_MER_ENN_10ÅR,
@@ -17,7 +17,7 @@ enum class MedlemskapUnntakÅrsak(override val resultat: Resultat = Resultat.OPP
     ANDRE_FORELDER_MEDLEM_MINST_5_ÅR_AVBRUDD_MINDRE_ENN_10_ÅR,
     ANDRE_FORELDER_MEDLEM_MINST_7_ÅR_AVBRUDD_MER_ENN_10_ÅR,
     TOTALVURDERING_OPPFYLLER_FORSKRIFT,
-    NEI(resultat = Resultat.IKKE_OPPFYLT)
+    NEI(mapping = JaNei.NEI)
 }
 
 class ForutgåendeMedlemskap
