+++ src/main/kotlin/no/nav/familie/ef/sak/regler/vilkårsregel/MorEllerFar.kt
@@ -3,7 +3,7 @@ package no.nav.familie.ef.sak.regler
 private enum class MorEllerFarRegel(override val id: String,
                                     override val beskrivelse: String) : RegelIdMedBeskrivelse {
 
-    OMSORG_FOR_EGNE_ELLER_ADOPTERTE_BARN("O1", "Har bruker omsorgen for egne/adopterte barn?"),
+    OMSORG_FOR_EGNE_ELLER_ADOPTERTE_BARN("M1", "Har bruker omsorgen for egne/adopterte barn?"),
 }
 
 class MorEllerFar : Vilkårsregel(vilkårType = VilkårType.MOR_ELLER_FAR,
